another outrageously minor bugfix in centos.h
Sorry! :grimacing:
The issue is, while `sed -i ' ' ...` is a perfectly valid syntax, it throws an error under Debian because for some reason the `sed` inside `Debian` expects this to be `sed -i' '...` (no whitespace after the `-i` switch) otherwise it says:
```
sed: can't read /^\s*$/d: No such file or directory
```
As expected, the solution is to remove the whitespace after `-i` in line #12. I am pretty sure this won't cause problems with other distros or other versions of sed.